"reverttt"

+++ train.py
@@ -30,9 +30,9 @@ def train(args):
     domA_train = CustomDataset(os.path.join(args.root, 'trainA.txt'), transform=comp_transform)
     domB_train = CustomDataset(os.path.join(args.root, 'trainB.txt'), transform=comp_transform)
 
-    A_label = torch.full((args.bs,), 1)
-    B_label = torch.full((args.bs,), 0)
-    B_separate = torch.full((args.bs, args.sep * (args.resize // 64) * (args.resize // 64)), 0)
+    A_label = torch.full((args.bs,), 1, dtype=torch.float)
+    B_label = torch.full((args.bs,), 0, dtype=torch.float)
+    B_separate = torch.full((args.bs, args.sep * (args.resize // 64) * (args.resize // 64)), 0, dtype=torch.float)
 
     e1 = E1(args.sep, args.resize // 64)
     e2 = E2(args.sep, args.resize // 64)
